Remove obsolete files
