+++ front/app/containers/Admin/projects/project/permissions/containers/Project/Granular/Timeline.tsx
@@ -10,7 +10,6 @@ import messages from 'components/admin/ActionsForm/messages';
 import { FormattedMessage } from 'utils/cl-intl';
 
 import PhasePermissions from '../../../components/PhasePermissions';
-import PhasePermissionsNew from '../../../components/PhasePermissionsNew';
 
 interface Props {
   projectId: string;
@@ -28,27 +27,14 @@ const Timeline = ({ projectId }: Props) => {
     <Box mb="20px">
       {phases.data &&
         phases.data.length > 0 &&
-        phases.data.map((phase, i) => {
-          if (isCustomPermittedByEnabled) {
-            return (
-              <PhasePermissionsNew
-                project={project.data}
-                phase={phase}
-                key={phase.id}
-                phaseNumber={i + 1}
-              />
-            );
-          }
-
-          return (
-            <PhasePermissions
-              project={project.data}
-              phase={phase}
-              key={phase.id}
-              phaseNumber={i + 1}
-            />
-          );
-        })}
+        phases.data.map((phase, i) => (
+          <PhasePermissions
+            project={project.data}
+            phase={phase}
+            key={phase.id}
+            phaseNumber={i + 1}
+          />
+        ))}
       {!phases.data ||
         (phases.data.length < 1 && (
           <p>
